Kill bogus isConsistent property, add a good one
When writing isConsistent I'd mis-remembered the details of how the
package worked. The property assumed that errored results would have
negative quotas, but in fact the quota is never lowered below zero; the
computation is stopped *before* state changes are made. The upshot of
this is that the test as intended didn't really make sense.

However, this brings to light an important property -- the quota is
never reduced below zero. So we have quickcheck test this now.